LayoutTests/media/media-source/media-source-timestampoffset-then-zero-expected.txt


EXPECTED (source.readyState == 'closed') OK
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer("video/mock; codecs=mock"))
RUN(sourceBuffer.timestampOffset = 5)
RUN(sourceBuffer.appendBuffer(makeSimpleAudio(0, 20, 0)))
EVENT(updateend)
After first append:
{PTS({5/1 = 5.000000}), DTS({5/1 = 5.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
{PTS({6/1 = 6.000000}), DTS({6/1 = 6.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
{PTS({7/1 = 7.000000}), DTS({7/1 = 7.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
{PTS({8/1 = 8.000000}), DTS({8/1 = 8.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
{PTS({9/1 = 9.000000}), DTS({9/1 = 9.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
{PTS({10/1 = 10.000000}), DTS({10/1 = 10.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
{PTS({11/1 = 11.000000}), DTS({11/1 = 11.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
{PTS({12/1 = 12.000000}), DTS({12/1 = 12.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
{PTS({13/1 = 13.000000}), DTS({13/1 = 13.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
{PTS({14/1 = 14.000000}), DTS({14/1 = 14.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
{PTS({15/1 = 15.000000}), DTS({15/1 = 15.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
{PTS({16/1 = 16.000000}), DTS({16/1 = 16.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
{PTS({17/1 = 17.000000}), DTS({17/1 = 17.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
{PTS({18/1 = 18.000000}), DTS({18/1 = 18.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
{PTS({19/1 = 19.000000}), DTS({19/1 = 19.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
{PTS({20/1 = 20.000000}), DTS({20/1 = 20.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
{PTS({21/1 = 21.000000}), DTS({21/1 = 21.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
{PTS({22/1 = 22.000000}), DTS({22/1 = 22.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
{PTS({23/1 = 23.000000}), DTS({23/1 = 23.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
{PTS({24/1 = 24.000000}), DTS({24/1 = 24.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
RUN(sourceBuffer.timestampOffset = 0)
RUN(sourceBuffer.appendBuffer(makeSimpleAudio(0, 20, 1)))
EVENT(updateend)
After second append:
{PTS({0/1 = 0.000000}), DTS({0/1 = 0.000000}), duration({1/1 = 1.000000}), flags(1), generation(1)}
{PTS({1/1 = 1.000000}), DTS({1/1 = 1.000000}), duration({1/1 = 1.000000}), flags(1), generation(1)}
{PTS({2/1 = 2.000000}), DTS({2/1 = 2.000000}), duration({1/1 = 1.000000}), flags(1), generation(1)}
{PTS({3/1 = 3.000000}), DTS({3/1 = 3.000000}), duration({1/1 = 1.000000}), flags(1), generation(1)}
{PTS({4/1 = 4.000000}), DTS({4/1 = 4.000000}), duration({1/1 = 1.000000}), flags(1), generation(1)}
{PTS({5/1 = 5.000000}), DTS({5/1 = 5.000000}), duration({1/1 = 1.000000}), flags(1), generation(1)}
{PTS({6/1 = 6.000000}), DTS({6/1 = 6.000000}), duration({1/1 = 1.000000}), flags(1), generation(1)}
{PTS({7/1 = 7.000000}), DTS({7/1 = 7.000000}), duration({1/1 = 1.000000}), flags(1), generation(1)}
{PTS({8/1 = 8.000000}), DTS({8/1 = 8.000000}), duration({1/1 = 1.000000}), flags(1), generation(1)}
{PTS({9/1 = 9.000000}), DTS({9/1 = 9.000000}), duration({1/1 = 1.000000}), flags(1), generation(1)}
{PTS({10/1 = 10.000000}), DTS({10/1 = 10.000000}), duration({1/1 = 1.000000}), flags(1), generation(1)}
{PTS({11/1 = 11.000000}), DTS({11/1 = 11.000000}), duration({1/1 = 1.000000}), flags(1), generation(1)}
{PTS({12/1 = 12.000000}), DTS({12/1 = 12.000000}), duration({1/1 = 1.000000}), flags(1), generation(1)}
{PTS({13/1 = 13.000000}), DTS({13/1 = 13.000000}), duration({1/1 = 1.000000}), flags(1), generation(1)}
{PTS({14/1 = 14.000000}), DTS({14/1 = 14.000000}), duration({1/1 = 1.000000}), flags(1), generation(1)}
{PTS({15/1 = 15.000000}), DTS({15/1 = 15.000000}), duration({1/1 = 1.000000}), flags(1), generation(1)}
{PTS({16/1 = 16.000000}), DTS({16/1 = 16.000000}), duration({1/1 = 1.000000}), flags(1), generation(1)}
{PTS({17/1 = 17.000000}), DTS({17/1 = 17.000000}), duration({1/1 = 1.000000}), flags(1), generation(1)}
{PTS({18/1 = 18.000000}), DTS({18/1 = 18.000000}), duration({1/1 = 1.000000}), flags(1), generation(1)}
{PTS({19/1 = 19.000000}), DTS({19/1 = 19.000000}), duration({1/1 = 1.000000}), flags(1), generation(1)}
{PTS({20/1 = 20.000000}), DTS({20/1 = 20.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
{PTS({21/1 = 21.000000}), DTS({21/1 = 21.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
{PTS({22/1 = 22.000000}), DTS({22/1 = 22.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
{PTS({23/1 = 23.000000}), DTS({23/1 = 23.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
{PTS({24/1 = 24.000000}), DTS({24/1 = 24.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
END OF TEST
